fix navigation by id issue
After clicking a navigation link and navigating to the page section by id, the fixed navigation bar would partially cover the section, hiding the heading, which is the most important indicator that the user has gotten to the correct spot. Adding padding to the top of the containers and removing whitespace elsewhere fixed this.